[maint] Rename trial_response dir to response
